feat: ConsoleBlackjackBame을 통한 Application 의존 최소화
